it compiles
i removed ScopedFIle because Data reader/writer are intended to be observe not own the memory.
Now must reimplement asio support from several months ago or another esp-specific solution. I could write a Valheim mod to connect directly to the esp server but that might be too complicated.
It might be easier to create a proxy on Valhalla on my main pc to receive steam packets directly from Valheim then forward those unencrypted over tcp/wifi to the esp.
got some testing and thinking to do in the meantime